python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermes-news"
version = "0.1.0"
description = "Self-hosted information agent — fetches, tags, scores, and filters internet content"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "feedparser>=6.0",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
hermes = "hermes_news.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hermes_news"]

[tool.ruff]
target-version = "py311"
line-length = 100
